WFの変更、テストの修正

+++ cmd/maadm/maadm_test.go
@@ -104,7 +104,7 @@ var _ = Describe("Marmotd Test", Ordered, func() {
 			Expect(err).NotTo(HaveOccurred())
 			Expect(h.NodeName).To(Equal("hvc"))
 			Expect(*h.IpAddr).To(Equal("127.0.0.1"))
-			Expect(h.Cpu).To(Equal(4))
+			Expect(h.Cpu).To(Equal(int32(4)))
 			Expect(*h.Memory).To(Equal(16384))
 			Expect(*(*h.StgPool)[0].VolGroup).To(Equal("vg1"))
 			Expect(*(*h.StgPool)[1].VolGroup).To(Equal("vg2"))
